[Refactor] Recipe Page Refactoring
