+++ nengboo_web/components/SubHeader.tsx
@@ -1,0 +1,13 @@
+export const SubHeader = (title: { title: string }) => {
+  return (
+    <div className="flex items-center justify-center mb-8">
+      <div className="w-[140px] h-9 p-2 bg-neutral-100 rounded-[100px] justify-center items-center gap-2.5 inline-flex mt-32">
+        <div className="w-[113px] h-[22px] justify-center items-center flex">
+          <p className="w-[113px] h-[22px] text-center text-neutral-800 text-xl font-medium ">
+            {title.title}
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+};
